moved to workspace, to have both examples working in parallel
