+++ examples/ble/esp32c6/.clippy.toml
