+++ examples/ble/esp32c6/Cargo.toml
@@ -1,0 +1,74 @@
+[package]
+edition = "2024"
+name = "esp32c6_example_with_ble"
+
+[dependencies]
+### Dependencies for all architectures:
+embassy-executor = { version = "0.9.1", features = [
+  "executor-thread",
+  "defmt",
+] }
+embassy-time = { version = "0.5.0", features = [
+  "defmt",
+  "defmt-timestamp-uptime",
+] }
+embassy-sync = { version = "0.7.2", features = ["defmt"] }
+embassy-futures = { version = "0", features = ["defmt"] }
+embassy-embedded-hal = "0.5.0"
+
+embedded-hal = { version = "1.0.0" }
+embedded-hal-async = { version = "1.0.0" }
+embedded-hal-bus = { version = "0.3.0", features = ["async"] }
+
+# For logging
+defmt = "1.0.1"
+defmt-rtt = "1.1.0"
+rtt-target = { version = "0.6.2", features = ["defmt"] }
+
+# panic handling
+critical-section = "1.2.0"
+panic-rtt-target = { version = "0.2.0", features = ["defmt"] }
+static_cell = "2.1.1"
+
+# For the messages sent around
+serde = { version = "1.0.228", features = ["derive"], default-features = false }
+postcard = "1.1.3"
+
+### ---- to use with the ESP32-C6  ----
+esp-hal = { version = "~1.0", features = ["defmt", "esp32c6", "unstable"] }
+esp-bootloader-esp-idf = { version = "0.4.0", features = ["defmt", "esp32c6"] }
+esp-rtos = { version = "0.2.0", features = ["embassy", "esp32c6", "esp-radio"] }
+# For the C6 dev board smart led
+esp-hal-smartled = { version = "0.17.0", features = ["esp32c6"] }
+smart-leds = { version = "0.4.0" }
+smart-leds-trait = { version = "0.3.2" }
+
+
+### --- for BLE  ----
+esp-radio = { version = "0.17.0", features = [
+  "esp32c6",
+  "ble",
+  "unstable",
+  "defmt",
+] }
+trouble-host = { version = "0.5.1", features = [
+  "default-packet-pool-mtu-255",
+  "derive",
+  "defmt",
+] }
+esp-alloc = { version = "0.9.0", features = ["esp32c6"] }
+
+
+[profile.dev]
+# Rust debug is too slow.
+# For debug builds always builds with some optimization
+opt-level = 3 # was "s", chose 2 for esp-radio
+
+[profile.release]
+codegen-units = 1        # LLVM can perform better optimizations using a single thread
+debug = 2
+debug-assertions = false
+incremental = false
+lto = 'fat'
+opt-level = 's'
+overflow-checks = false
