+++ examples/ble/esp32c6/build.rs
