+++ examples/ble/esp32c6/src/bas_peripheral.rs
