+++ examples/lora/rak3272s/Cargo.toml
@@ -1,0 +1,49 @@
+[package]
+edition = "2024"
+name = "must-stm32wle5cx-examples"
+version = "0.1.0"
+
+[dependencies]
+# Change stm32wle5cc to your chip name, if necessary. Also update .cargo/config.toml
+embassy-stm32 = { version = "0.5.0", features = [
+  "defmt",
+  "stm32wle5cc",
+  "time-driver-any",
+  "memory-x",
+  "unstable-pac",
+  "exti",
+  "chrono",
+] }
+embassy-executor = { version = "0.9.1", features = [
+  "arch-cortex-m",
+  "executor-thread",
+  "defmt",
+] }
+embassy-time = { version = "0.5.0", features = [
+  "defmt",
+  "defmt-timestamp-uptime",
+] }
+embassy-sync = { version = "0.7.2", features = ["defmt"] }
+embassy-futures = { version = "0", features = ["defmt"] }
+
+# lora-phy = { version = "3.0.1", features = ["lorawan-radio", "defmt-03"] }
+lora-phy = { git = "https://github.com/lora-rs/lora-rs.git", features = [
+  "defmt-03",
+] }
+
+defmt = "1.0.1"
+defmt-rtt = "1.1.0"
+panic-probe = { version = "1.0.0", features = ["print-defmt"] }
+
+cortex-m = { version = "0.7.6", features = [
+  "inline-asm",
+  "critical-section-single-core",
+] }
+cortex-m-rt = "0.7.0"
+embedded-hal = { version = "1.0.0" }
+embedded-hal-async = { version = "1.0.0" }
+embedded-hal-bus = { version = "0.3.0", features = ["async"] }
+
+must-hop = { workspace = true }
+[profile.release]
+debug = 2
