+++ examples/lora/rak3272s/src/bin/lora_p2p_send.rs
