+++ must-hop/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "must-hop"
+version = "0.1.0"
+edition = "2024"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+
+defmt = "1.0.1"
