+++ must-hop/src/lib.rs
@@ -1,0 +1,5 @@
+use defmt::info;
+
+pub fn hello_world() {
+    info!("Hello world from must hop!");
+}
