--- src/lib.rs
@@ -1,5 +1,0 @@
-#![no_std]
-
-pub mod bas_peripheral;
-pub mod led_runner;
-pub mod lora_tasks;
